Changes to make sure quarkiverse logging-manager can reuse the static content Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
