+++ extensions/vertx-http/deployment/src/main/resources/dev-templates/logmanagerLog.html
@@ -1,0 +1,8 @@
+<div id="logstreamManagerLog">
+    <a id="logstreamInformationSection"></a>
+    <a id="logstreamFollowLogButton" data-toggle="tooltip" data-placement="left" title="Follow log (auto scroll)"><i id="logstreamFollowLogButtonIcon" class="text-success fas fa-circle-notch fa-spin"></i></a>
+    <div id="logstreamLogTerminal" class="bg-dark text-light ">
+        <samp id="logstreamLogTerminalText" class="text-wrap">
+        </samp>
+    </div>
+</div>
