+++ extensions/vertx-http/runtime/src/main/java/io/quarkus/vertx/http/runtime/logstream/LogStreamWebSocket.java
@@ -1,21 +1,7 @@
 package io.quarkus.vertx.http.runtime.logstream;
 
-import static java.util.logging.Level.ALL;
-import static java.util.logging.Level.CONFIG;
-import static java.util.logging.Level.FINE;
-import static java.util.logging.Level.FINER;
-import static java.util.logging.Level.FINEST;
-import static java.util.logging.Level.INFO;
-import static java.util.logging.Level.OFF;
-import static java.util.logging.Level.SEVERE;
-import static java.util.logging.Level.WARNING;
-
-import java.util.ArrayList;
-import java.util.Enumeration;
 import java.util.List;
-import java.util.TreeMap;
 import java.util.logging.Level;
-import java.util.logging.LogManager;
 import java.util.logging.Logger;
 
 import org.jboss.logmanager.ExtHandler;
@@ -26,7 +12,6 @@ import io.netty.handler.codec.http.HttpHeaderNames;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.http.ServerWebSocket;
-import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.RoutingContext;
 
@@ -42,6 +27,8 @@ public class LogStreamWebSocket implements Handler<RoutingContext> {
     private final ExtHandler rootHandler;
     private final org.jboss.logmanager.Logger rootLogger;
 
+    private final LogController logController = new LogController();
+
     public LogStreamWebSocket() {
 
         // Add history handler
@@ -51,8 +38,8 @@ public class LogStreamWebSocket implements Handler<RoutingContext> {
         addHandler(historyHandler);
 
         initMessage.put(TYPE, INIT);
-        initMessage.put("loggers", getLoggers());
-        initMessage.put("levels", getLevels());
+        initMessage.put("loggers", logController.getLoggers());
+        initMessage.put("levels", logController.getLevels());
     }
 
     @Override
@@ -130,73 +117,12 @@ public class LogStreamWebSocket implements Handler<RoutingContext> {
         session.started = false;
     }
 
-    private JsonArray getLevels() {
-        return new JsonArray()
-                .add(OFF.getName())
-                .add(SEVERE.getName())
-                .add(WARNING.getName())
-                .add(INFO.getName())
-                .add(CONFIG.getName())
-                .add(FINE.getName())
-                .add(FINER.getName())
-                .add(FINEST.getName())
-                .add(ALL.getName());
-    }
-
-    private JsonArray getLoggers() {
-        TreeMap<String, JsonObject> loggerMap = new TreeMap<>();
-        LogManager manager = LogManager.getLogManager();
-        Enumeration<String> loggerNames = manager.getLoggerNames();
-        while (loggerNames.hasMoreElements()) {
-            String loggerName = loggerNames.nextElement();
-            JsonObject jsonObject = getLogger(loggerName);
-            if (jsonObject != null) {
-                loggerMap.put(loggerName, jsonObject);
-            }
-        }
-
-        List<JsonObject> orderedLoggers = new ArrayList<>(loggerMap.values());
-        JsonArray jsonArray = new JsonArray(orderedLoggers);
-        return jsonArray;
-    }
-
-    private JsonObject getLogger(String loggerName) {
-        if (loggerName != null && !loggerName.isEmpty()) {
-            Logger logger = Logger.getLogger(loggerName);
-            JsonObject jsonObject = new JsonObject();
-            jsonObject.put("name", loggerName);
-            jsonObject.put("effectiveLevel", getEffectiveLogLevel(logger));
-            jsonObject.put("configuredLevel", getConfiguredLogLevel(logger));
-            return jsonObject;
-        }
-        return null;
-    }
-
-    private String getConfiguredLogLevel(Logger logger) {
-        Level level = logger.getLevel();
-        return level != null ? level.getName() : null;
-    }
-
-    private String getEffectiveLogLevel(Logger logger) {
-        if (logger == null) {
-            return null;
-        }
-        if (logger.getLevel() != null) {
-            return logger.getLevel().getName();
-        }
-        return getEffectiveLogLevel(logger.getParent());
-    }
-
     private void update(String message) {
         String[] p = message.split("\\|");
         if (p.length == 3) {
             String loggerName = p[1];
             String levelVal = p[2];
-            Logger logger = Logger.getLogger(loggerName);
-            if (logger != null) {
-                Level level = Level.parse(levelVal);
-                logger.setLevel(level);
-            }
+            logController.updateLogLevel(loggerName, levelVal);
         }
     }
